ispravke

+++ src/main/java/com/ftn/ZgradeProjekat/repository/SurveyRepository.java
@@ -12,6 +12,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 
 
-public interface SurveyRepository extends JpaRepository<Survey,Integer> {
+public interface SurveyRepository extends JpaRepository<Survey,Long> {
 
 }
